coco/support meta.json for fun local and fun package
