+++ lib/commands/deploy.js
@@ -1,32 +1,8 @@
 'use strict';
-const fs = require('fs-extra');
-const path = require('path');
 
 const { red } = require('colors');
-const { readJsonFromFile } = require('../utils/file');
-const { getRootTplPath, detectTplPath, validateTplName, DEFAULT_BUILD_ARTIFACTS_PATH_SUFFIX} = require('../tpl');
-
-const _ = require('lodash');
-
-async function getModifiedTimes(tplPath) {
-  if (tplPath.indexOf(DEFAULT_BUILD_ARTIFACTS_PATH_SUFFIX) === -1) { return {}; }
-
-  const rootTplPath = getRootTplPath(tplPath);
-  const metaPath = path.resolve(path.dirname(rootTplPath), DEFAULT_BUILD_ARTIFACTS_PATH_SUFFIX, 'meta.json');
-
-  if (!await fs.pathExists(metaPath)) { return {}; }
-
-  const metaObj = await readJsonFromFile(metaPath);
-
-  if (_.isEmpty(metaObj)) {
-    return {};
-  }
-
-  return _.pickBy((metaObj.modifiedTimes || {}), (mtime, filePath) => {
-    const lstat = fs.lstatSync(filePath);
-    return mtime !== lstat.mtime.getTime().toString();
-  });
-}
+const { ensureFilesModified } = require('../utils/file');
+const { detectTplPath, validateTplName } = require('../tpl');
 
 async function deploy(context) {
   let tplPath = context.template;
@@ -41,14 +17,7 @@ async function deploy(context) {
 
   validateTplName(tplPath);
 
-  const modifiedTimes = await getModifiedTimes(tplPath); // for fun build scene
-
-  if (!_.isEmpty(modifiedTimes)) {
-    throw new Error(`
-        ${Object.keys(modifiedTimes).join('\n\t')}\n` +
-`
-Fun detected the above path have been modified. Please execute ‘fun build’ to compile your functions.`);
-  }
+  await ensureFilesModified(tplPath);
 
   await require('../deploy/deploy-by-tpl').deploy(tplPath, context);
 }
