+++ lib/commands/local/invoke.js
@@ -5,10 +5,10 @@ const path = require('path');
 const debug = require('debug')('fun:local');
 const definition = require('../../definition');
 
-const { getEvent } = require('../../utils/file');
 const { red, yellow } = require('colors');
 const { ensureTmpDir } = require('../../utils/path');
 const { getDebugPort, getDebugIde } = require('../../debug');
+const { getEvent, ensureFilesModified } = require('../../utils/file');
 const { generateMergedTpl, detectNasBaseDir } = require('../../tpl');
 
 async function localInvoke(invokeName, tplPath, tpl, debugPort, event, debugIde, baseDir, tmpDir, debuggerPath, debugArgs, reuse = true, nasBaseDir) {
@@ -47,7 +47,7 @@ async function invoke(invokeName, options) {
 
   const event = await getEvent(options.event);
   debug('event content: ' + event);
-  
+
   const debugPort = getDebugPort(options);
   const debugIde = getDebugIde(options);
 
@@ -56,8 +56,10 @@ async function invoke(invokeName, options) {
 
   const baseDir = path.resolve(path.dirname(tplPath));
   const reuse = options.reuse;
-  
+
   const nasBaseDir = detectNasBaseDir(tplPath);
+
+  await ensureFilesModified(tplPath);
 
   await localInvoke(invokeName, tplPath, tpl, debugPort, event, debugIde, baseDir, options.tmpDir, debuggerPath, debugArgs, reuse, nasBaseDir);
 }
