+++ lib/commands/local/start.js
@@ -10,6 +10,7 @@ const httpSupport = require('./http-support');
 const EventStart = require('../../local/event-start');
 
 const { ensureTmpDir } = require('../../utils/path');
+const { ensureFilesModified } = require('../../utils/file');
 const { showLocalStartNextTips } = require('../../../lib/build/tips');
 const { promptForDebugaHttptriggers } = require('../../../lib/init/prompt');
 const { getDebugPort, getDebugIde } = require('../../debug');
@@ -167,6 +168,8 @@ async function start(options, invokeName = '') {
   const debuggerPath = options.debuggerPath;
   const debugArgs = options.debugArgs;
   const baseDir = path.dirname(tplPath);
+
+  await ensureFilesModified(tplPath);
 
   let httpTriggers = definition.findHttpTriggersInTpl(tpl);
