+++ lib/package/package.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const { ensureFilesModified } = require('../utils/file');
 const { isEmptyDir } = require('../nas/cp/file');
 const { getOssClient } = require('../client');
 const { parseMountDirPrefix } = require('../fc');
@@ -168,6 +169,9 @@ async function pack(tplPath, bucket, outputTemplateFile) {
   if (!bucket) {
     throw new Error('Missing OSS bucket');
   }
+
+  await ensureFilesModified(tplPath);
+
   const ossClient = await getOssClient(bucket);
 
   const updatedEnvTpl = await processNasPythonPaths(tpl, tplPath);
