+++ lib/utils/file.js
@@ -1,10 +1,14 @@
 'use strict';
 
 const fs = require('fs-extra');
+const path = require('path');
 const readline = require('readline');
 const getStdin = require('get-stdin');
 
 const { red } = require('colors');
+const { getRootTplPath, DEFAULT_BUILD_ARTIFACTS_PATH_SUFFIX } = require('../tpl');
+
+const _ = require('lodash');
 
 const getVisitor = require('../visitor').getVisitor;
 
@@ -64,13 +68,13 @@ async function getEvent(eventFile, ec = 'local invoke', dp = '/fun/local/invoke'
           el: 'success',
           dp
         }).send();
-  
+
         resolve(event);
       });
     });
-    
+
     rl.on('SIGINT', function () {
-      
+
       getVisitor().then(visitor => {
         visitor.event({
           ec,
@@ -78,7 +82,6 @@ async function getEvent(eventFile, ec = 'local invoke', dp = '/fun/local/invoke'
           el: 'cancel',
           dp
         }).send();
-  
         // Keep the behavior consistent with system.
         reject(new Error('^C'));
       });
@@ -132,4 +135,36 @@ async function readJsonFromFile(absFilePath) {
   return obj;
 }
 
-module.exports = { readLines, getEvent, recordMtimes, readJsonFromFile };
+async function getModifiedTimestamps(tplPath) {
+  if (tplPath.indexOf(DEFAULT_BUILD_ARTIFACTS_PATH_SUFFIX) === -1) { return {}; }
+
+  const rootTplPath = getRootTplPath(tplPath);
+  const metaPath = path.resolve(path.dirname(rootTplPath), DEFAULT_BUILD_ARTIFACTS_PATH_SUFFIX, 'meta.json');
+
+  if (!await fs.pathExists(metaPath)) { return {}; }
+
+  const metaObj = await readJsonFromFile(metaPath);
+
+  if (_.isEmpty(metaObj)) {
+    return {};
+  }
+
+  return _.pickBy((metaObj.modifiedTimes || {}), (mtime, filePath) => {
+    const lstat = fs.lstatSync(filePath);
+    return mtime !== lstat.mtime.getTime().toString();
+  });
+}
+
+// for fun build scene
+async function ensureFilesModified(tplPath) {
+  const modifiedTimes = await getModifiedTimestamps(tplPath);
+
+  if (!_.isEmpty(modifiedTimes)) {
+    throw new Error(`
+        ${Object.keys(modifiedTimes).join('\n\t')}\n` +
+`
+Fun detected the above path have been modified. Please execute ‘fun build’ to compile your functions.`);
+  }
+}
+
+module.exports = { readLines, getEvent, recordMtimes, readJsonFromFile, ensureFilesModified };
